haha yes
Updated Patchouli
Sorting now has a patchouli page type
Moved client registry classes to `registry.client`
Renamed some classes
Fixed Rocket spamming log because write and read got switched lol
