+++ gradle.properties
@@ -11,7 +11,7 @@ api_version = 0.14.1+build.372-1.16
 
 # Mod
 mod_name = astromine
-mod_version = 1.0.18
+mod_version = 1.0.19
 mod_group = com.github.chainmailstudios
 version_meta = fabric-1.16.1
 
@@ -19,7 +19,7 @@ version_meta = fabric-1.16.1
 cardinal_components_version = 2.4.1
 spinnery_version = 3.0.48+fabric-1.16.x
 shapes_version = 2.0.0+build.10
-patchouli_version = 1.16-37.20-FABRIC
+patchouli_version = 1.16-37.21-FABRIC
 lba_version=0.7.0
 
 # Conveniences
