+++ src/main/java/com/github/chainmailstudios/astromine/client/patchouli/SortingPage.java
@@ -1,0 +1,10 @@
+package com.github.chainmailstudios.astromine.client.patchouli;
+
+import com.github.chainmailstudios.astromine.common.recipe.SortingRecipe;
+import vazkii.patchouli.client.book.page.abstr.PageSimpleProcessingRecipe;
+
+public class SortingPage extends PageSimpleProcessingRecipe<SortingRecipe> {
+    public SortingPage() {
+        super(SortingRecipe.Type.INSTANCE);
+    }
+}
