+++ src/main/java/com/github/chainmailstudios/astromine/common/recipe/LiquidGeneratingRecipe.java
@@ -23,6 +23,7 @@ import com.github.chainmailstudios.astromine.common.fraction.Fraction;
 import com.github.chainmailstudios.astromine.common.utilities.FractionUtilities;
 import com.github.chainmailstudios.astromine.common.volume.energy.EnergyVolume;
 import com.github.chainmailstudios.astromine.common.volume.fluid.FluidVolume;
+import com.github.chainmailstudios.astromine.registry.AstromineBlocks;
 import com.github.chainmailstudios.astromine.registry.AstromineComponentTypes;
 
 import com.google.gson.Gson;
@@ -100,7 +101,12 @@ public class LiquidGeneratingRecipe implements Recipe<Inventory> {
 	public DefaultedList<Ingredient> getPreviewInputs() {
 		return DefaultedList.of(); // we are not dealing with items
 	}
-	
+
+	@Override
+	public ItemStack getRecipeKindIcon() {
+		return new ItemStack(AstromineBlocks.LIQUID_GENERATOR);
+	}
+
 	public Fluid getFluid() {
 		return fluid.get();
 	}
