+++ src/main/java/com/github/chainmailstudios/astromine/registry/client/AstromineBlockEntityRenderers.java
@@ -1,4 +1,4 @@
-package com.github.chainmailstudios.astromine.registry;
+package com.github.chainmailstudios.astromine.registry.client;
 
 import net.fabricmc.fabric.api.client.rendereregistry.v1.BlockEntityRendererRegistry;
 
@@ -8,6 +8,7 @@ import net.minecraft.client.render.block.entity.BlockEntityRenderDispatcher;
 import net.minecraft.client.render.block.entity.BlockEntityRenderer;
 
 import com.github.chainmailstudios.astromine.client.render.block.HolographicBridgeBlockEntityRenderer;
+import com.github.chainmailstudios.astromine.registry.AstromineBlockEntityTypes;
 
 import java.util.function.Function;
