+++ src/main/java/com/github/chainmailstudios/astromine/registry/client/AstromineClientModels.java
@@ -1,4 +1,4 @@
-package com.github.chainmailstudios.astromine.registry;
+package com.github.chainmailstudios.astromine.registry.client;
 
 import com.github.chainmailstudios.astromine.AstromineCommon;
 import com.github.chainmailstudios.astromine.client.model.RocketEntityModel;
