+++ src/main/java/com/github/chainmailstudios/astromine/registry/client/AstromineClientPackets.java
@@ -1,4 +1,4 @@
-package com.github.chainmailstudios.astromine.registry;
+package com.github.chainmailstudios.astromine.registry.client;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -9,6 +9,8 @@ import net.minecraft.util.Identifier;
 
 import com.github.chainmailstudios.astromine.AstromineCommon;
 import com.github.chainmailstudios.astromine.common.entity.RocketEntity;
+import com.github.chainmailstudios.astromine.registry.AstromineCommonPackets;
+import com.github.chainmailstudios.astromine.registry.AstromineEntityTypes;
 
 import java.util.UUID;
