+++ src/main/java/com/github/chainmailstudios/astromine/registry/client/AstromineParticleFactories.java
@@ -1,6 +1,9 @@
-package com.github.chainmailstudios.astromine.registry;
+package com.github.chainmailstudios.astromine.registry.client;
 
 import com.github.chainmailstudios.astromine.client.particle.RocketFlameParticle;
+import com.github.chainmailstudios.astromine.registry.AstromineItems;
+import com.github.chainmailstudios.astromine.registry.AstromineParticles;
+
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
