+++ src/main/resources/data/astromine/patchouli_books/manual/en_us/entries/world/ore_clusters.json
@@ -8,6 +8,12 @@
 			"type": "text",
 			"title": "text.astromine.world.ore_clusters.page_one.title",
 			"text": "text.astromine.world.ore_clusters.page_one.text"
+		},
+		{
+			"type": "astromine:sorting",
+			"recipe": "astromine:metite_dust_from_sorting_cluster",
+			"title": "text.astromine.world.ore_clusters.page_two.title",
+			"text": "text.astromine.world.ore_clusters.page_two.text"
 		}
 	]
 }
